+++ toys-core/lib/toys/standard_middleware/show_help.rb
@@ -188,7 +188,7 @@ module Toys
 
       def output_help(str)
         if less_path
-          Utils::Exec.new.exec([less_path, "-R"], in_from: str)
+          Utils::Exec.new.exec([less_path, "-R"], in: str)
         else
           terminal.puts(str)
         end