+++ toys-core/CHANGELOG.md
@@ -4,7 +4,9 @@
 
 * CHANGED: ModuleLookup is now a customizable class and can have multiple sources.
 * CHANGED: Moved the existing templates to the toys gem since they are rake replacements.
+* CHANGED: Renamed :in_from, :out_to, and :err_to exec options to :in, :out, :err
 * ADDED: CLI can now customize the standard helpers, templates, and middleware.
+* IMPROVED: Exec raises an error if passed an unknown option.
 
 ### 0.3.7.1 / 2018-05-30