+++ toys/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### 0.3.8 / TBD
 
+* CHANGED: Renamed :in_from, :out_to, and :err_to exec options to :in, :out, :err
+
 ### 0.3.7.1 / 2018-05-30
 
 * FIXED: Fix crash in system update.